backend/pyproject.toml
[tool.poetry]
name = "Template project"
version = "0.1.0"
description = "template project"
license = "MIT"
authors=["Egor Ponyavin <ep@idaproject.com>"]

[tool.poetry.dependencies]
# core
python = "3.12.*"
Django = "5.1.*"
psycopg2-binary = "2.9.*"
sentry-sdk = "1.39.*"
requests = "2.31.*"
boto3 = "1.20.*"
django-redis-cache = "3.0.*"
django-storages = "1.12.*"
gunicorn = "^20.1.0"
django-user-agents = "^0.4.0"

# Models \ Admin
django-solo = "2.0.*"
django-ckeditor = "6.2.*"
django-admin-sortable2 = "2.1.*"
django-location-field = "~2.1.0"
phonenumbers = "^8.12.29"
django-phonenumber-field = "5.2.*"
django-admin-list-filter-dropdown = "~1.0.3"

# Rest
djangorestframework = "3.15.*"
django-filter = "22.*"
drf-spectacular = "0.24.*"

# Test
factory-boy = "^3.2.0"
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"
pytest-sugar = "^0.9.4"

# Celery
celery = "5.4.0"
django-celery-email = "~3.0.0"

# Others
Pillow = "10.0.1"
django-robots = "5.0.*"
jsonschema = "^4.21.1"
django-jsonform = "^2.22.0"
json-log-formatter = "^1.0"
pytz = "^2024.2"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "4.4.*"

# linting
ruff = "0.0.*"

# Typehinting
mypy = "1.*"
mypy-extensions = "1.*"
django-stubs = "1.14.*"
djangorestframework-stubs="1.8.*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["ALL"]
ignore = [
    "ANN101",
    "A003",
    "RUF001",
    "D100",
    "RUF002",
    "ANN001",
    "ANN201",
    "RUF012",
    "D106",
    "ARG002",
    "ANN003",
    "ANN002",
    "D105",
    "EXE002",
    "A001",
    "ARG001",
    "BLE001",
    "PTH123",
    "PTH118",
    "C901",
    "PLR0912",
    "PLR0913",
    "FBT001",
    "FBT002",
    "SLF001",
    "RUF003",
    "B023",
    "PTH122",
    "S113",
    "S324",
    "S104",
    "FLY002",
    "PLR0915",
    "S106",
    "S105",
    "B018",
    "N806",
    "S314",
    "N817",
    "PTH110",
    "PTH103",
    "PTH109",
    "SIM103",
    "PT001",
]
src = ["."]
ignore-init-module-imports = true
target-version = "py312"

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "**/migrations/**",
    "__init__.py",
    "apps.py",
    "**/tests/**",
    "manage.py",
]

# Same as Black.
line-length = 100

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
